--- app/src/App.css
@@ -1,281 +1,0 @@
-.not-connected-page {
-  width: 100%;
-  height: 100%;
-
-}
-body {
-  background-color: rgb(245, 241, 237);
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-}
-.zankoocodeDEX {
-  margin-top: 2rem;
-  height: 10%;
-}
-.connect-button {
-  width: 15%;
-  height: 12%;
-  padding: 2rem;
-  padding-bottom: 3.4rem;
-}
-.image-div {
-  margin-top: 5rem;
-  margin-bottom: 5rem;
-  width: 25%;
-  height: 40%;
-}
-
-img {
-  width: 100%;
-  height: 100%;
-}
-
-
-button {
-  text-align: center;
-  background-color: rgb(63, 139, 223);
-  color: white;
-  border: none;
-  border-radius: 1rem;
-  font-size: 1.2rem;
-  width: 10%;
-  height: 10%;
-  padding: 1.4rem;
-  
-  padding-top: 1rem;
-  padding-bottom: 2rem;
-}
-button:hover {
-  cursor: pointer;
-}
-.loading-time-sec {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-.loading {
-  margin-top: 12rem;
-}
-.description{
-    width: 100%;
-    height: 10%;
-    margin-top: 4rem;
-    text-align: center; 
-}
-
-.liqudity-sec{
-  width: 100%;
-  height: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  
-  margin-top: 2rem;
-}
-
-input {
-  width: 20%;
-  height: 15%;
-  margin-top: 1rem;
-  padding: 1.5rem;
-  border-radius: 1rem;
-  border: 1.5px solid wheat;
-}
-
-
-.addLiquidity-button {
-  margin-top: 1rem;
-  padding-top: 1rem;
-  padding-bottom: 2rem;
-}
-
-.remove-liquidity-sec {
-  width: 100%;
-  height: 10%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 2rem;
-  text-align: center;
-}
-
-.input-div {
-  margin-top: 1rem;
-}
-
-.removeLiquidity {
-  margin-top: 2rem;
-  padding-top: 1rem;
-  padding-bottom: 2.4rem;
-}
-
-.swap-sec {
-  width: 100%;
-  height: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  margin-top: 8rem;
-}
-
-.token-select {
-  width: 20%;
-  height: 5%;
-  margin-top: 1rem;
-  padding: 1rem;
-  border-radius: 1rem;
-}
-.swap-button {
-  margin-top: 2rem;
-}
-.image-sec {
-  width: 20%;
-  height: 40%;
-  margin-bottom: 7rem;
-  margin-top: 2rem;
-}
-
-.App {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  
-  text-align: center;
-}
-
-.welcome {
-  margin-top: 3rem;
-  height: 10%;
-}
-.DEX-description {
-  height: 10%;
-  margin-top: 2rem;
-}
-
-
-
-
-.liquidity-tab {
-  
-  padding-top: 1.2rem;
-  padding-bottom: 2.4rem;
-}
-
-.swap-tab {
-  margin-top: 1rem;
-  margin-left: 1rem;
-  padding-top: 1.2rem;
-  padding-bottom: 2.4rem;
-}
-.addliq {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  height: 30%;
-  margin-bottom: 2rem;
-  margin-top: 2rem;
-}
-
-.ZCD-needed {
-  margin-top: 1rem;
-}
-
-
-@media only screen and (max-width: 500px) {
-  .connect-button {
-    width: 30%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 3.4rem;
-    margin-top: 2rem;
-  }
-  .image-div {
-    
-    width: 40%;
-    height: 48%;
-    margin-top: 4rem;
-    margin-bottom: 10rem;
-  }
-
-  .welcome {
-    font-size: 1.5rem;
-  }
-
-  .liquidity-tab {
-    width: 25%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-    margin-top: 2rem;
-  }
-
-  .swap-tab {
-    width: 20%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-    margin-top: 2rem;
-  }
-
-  .image-sec {
-    width: 40%;
-    height: 48%;
-    margin-top: 4rem;
-    margin-bottom: 10rem;
-  }
-
-  .input-ether-amount {
-    width: 30%;
-  }
-  .input-zankoocode-amount {
-    width: 30%;
-  }
-  .lp-tokens-input {
-    width: 30%;
-  }
-  .addLiquidity-button2 {
-    width: 20%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-  }
-  .addLiquidity-button {
-    width: 20%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-  }
-  .removeLiquidity {
-    width: 25%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-  }
-  .swap-input {
-    width: 30%;
-  }
-  .swap-button {
-    width: 20%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-  }
-
-  .loading {
-    width: 25%;
-    height: 10%;
-    padding: 1rem;
-    padding-bottom: 2rem;
-  }
-}
