+++ app/src/components/home.css
@@ -1,0 +1,92 @@
+.not-connected-page {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+  }
+  
+  .zankoocodeDEX {
+    text-align: center;
+    margin-top: 2rem;
+    height: 10%;
+  }
+  .walletConnect-btn {
+    margin-top: 2rem;
+  }
+  
+  button {
+    height: 10%;
+    padding: 1rem;
+    border-radius: 1rem;
+    border: none;
+    background-color: rgba(77, 157, 232, 1);
+    color: white;
+  }
+  button:hover{
+    cursor: pointer;
+  }
+  .image-div {
+    margin-top: 4rem;
+    width: 30%;
+    height: 20%;
+    margin-bottom: 4rem;
+  }
+  
+  img {
+    width: 100%;
+    height: 100%;
+  }
+  
+  
+  .App {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+  }
+  
+  .welcome {
+    margin-top: 3rem;
+  }
+  
+  .DEX-description {
+    margin-top: 1rem;
+  }
+  
+  .buttons {
+    margin-top: 2rem;
+    display: flex;
+    justify-content: space-around;
+    width: 20%;
+  }
+  
+  .liquidity-tab-btn  {
+    width: 40%;
+  }
+  .swap-tab-btn {
+    width: 40%;
+  }
+  .image-sec {
+    margin-top: 2rem;
+    margin-bottom: 2rem;
+  }
+  @media only screen and (max-width: 600px){
+    .image-div {
+      margin-top: 6rem;
+      width: 40%;
+      height: 30%;
+      margin-bottom: 10rem;
+    }
+    .buttons {
+      margin-top: 2rem;
+      display: flex;
+      justify-content: space-around;
+      width: 50%;
+    } 
+  }
