+++ app/src/index.css
@@ -1,10 +1,8 @@
-
 * {
-  width: 100%;
-  height: 100%;
-  margin: 0;
- 
-  box-sizing: border-box;
   
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: sans-serif;
   
 }
